完善PHPOK

+++ framework/model/email.php
@@ -1,14 +1,18 @@
 <?php
-/***********************************************************
-	Filename: {phpok}/model/email.php
-	Note	: 邮件内容管理器
-	Version : 3.0
-	Author  : qinggan
-	Update  : 2013年06月30日 23时42分
-***********************************************************/
+/**
+ * 邮件内容管理器
+ * @package phpok\model
+ * @作者 qinggan <admin@phpok.com>
+ * @版权 深圳市锟铻科技有限公司
+ * @主页 http://www.phpok.com
+ * @版本 4.x
+ * @授权 http://www.phpok.com/lgpl.html PHPOK开源授权协议：GNU Lesser General Public License
+ * @时间 2017年04月22日
+**/
+
 class email_model_base extends phpok_model
 {
-	function __construct()
+	public function __construct()
 	{
 		parent::model();
 	}
@@ -20,17 +24,16 @@ class email_model_base extends phpok_model
 	}
 
 
-	function get_one($id)
+	public function get_one($id)
 	{
-		if(!$id)
-		{
+		if(!$id){
 			return false;
 		}
 		$sql = "SELECT * FROM ".$this->db->prefix."email WHERE id='".$id."'";
 		return $this->db->get_one($sql);
 	}
 
-	function get_list($condition="",$offset=0,$psize=20)
+	public function get_list($condition="",$offset=0,$psize=20)
 	{
 		$sql = " SELECT * FROM ".$this->db->prefix."email ";
 		if($condition)
@@ -41,47 +44,59 @@ class email_model_base extends phpok_model
 		return $this->db->get_all($sql);
 	}
 
-	function simple_list($siteid=0)
+	public function simple_list($siteid=0)
 	{
 		$condition = $siteid ? "site_id IN(0,".$siteid.")" : "site_id=0";
-		$sql = "SELECT id,identifier,title FROM ".$this->db->prefix."email WHERE ".$condition;
+		$sql = "SELECT id,identifier,title,note FROM ".$this->db->prefix."email WHERE ".$condition;
 		return $this->db->get_all($sql);
 	}
 
 	//取得总数量
-	function get_count($condition="")
+	public function get_count($condition="")
 	{
 		$sql = "SELECT count(id) FROM ".$this->db->prefix."email ";
-		if($condition)
-		{
+		if($condition){
 			$sql .= " WHERE ".$condition;
 		}
 		return $this->db->count($sql);
 	}
 
-	//存储邮件内容信息
-	function save($data,$id=0)
+	/**
+	 * 存储邮件内容信息
+	 * @参数 $data 数组，要写入的数据
+	 * @参数 $id 大于0时表示更新
+	**/
+	public function save($data,$id=0)
 	{
-		if($id)
-		{
+		if($id){
 			$this->db->update_array($data,"email",array("id"=>$id));
 			return true;
-		}
-		else
-		{
+		}else{
 			$insert_id = $this->db->insert_array($data,"email");
 			return $insert_id;
 		}
 	}
 
-	//删除邮件内容
-	function del($id)
+	/**
+	 * 删除邮件内容
+	 * @参数 $id 要删除的邮件ID，多个ID用英文逗号隔开
+	**/
+	public function del($id=0)
 	{
+		if(!$id){
+			return false;
+		}
 		$sql = "DELETE FROM ".$this->db->prefix."email WHERE id IN(".$id.")";
 		return $this->db->query($sql);
 	}
 
-	function get_identifier($identifier,$site_id=0,$id=0)
+	/**
+	 * 检测标识是否存在
+	 * @参数 $identifier 标识
+	 * @参数 $site_id 站点ID
+	 * @参数 $id 不检查指定的ID
+	**/
+	public function get_identifier($identifier,$site_id=0,$id=0)
 	{
 		if(!$site_id){
 			$site_id = $this->site_id;
@@ -94,9 +109,13 @@ class email_model_base extends phpok_model
 		return $this->db->get_one($sql);
 	}
 
+	/**
+	 * 取得模板内容
+	 * @参数 $code 标识ID
+	 * @参数 $site_id 站点ID
+	**/
 	public function tpl($code,$site_id=0)
 	{
 		return $this->get_identifier($code,$site_id);
 	}
 }
-?>
